pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vishwamitra-dropout-commons"
version = "0.1.0"
description = "Vishwamitra — Mechanism design for educational commons. A multi-agent reinforcement learning OpenEnv environment that learns to rewrite the rules of educational collapse before it happens."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Rudra Bhaskar" },
]
keywords = [
    "reinforcement-learning",
    "multi-agent",
    "openenv",
    "mechanism-design",
    "game-theory",
    "education",
    "gymnasium",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openenv-core>=0.2.0",
    "gradio>=4.0.0",
    "plotly>=5.0.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "huggingface-hub>=0.19.0",
    "gymnasium>=0.29.0",
    "openai>=1.30.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.27",
]

[project.urls]
Homepage = "https://huggingface.co/spaces/rudra9439/vidya-meta-rl"
Repository = "https://github.com/RudraBhaskar9439/Enigma"
"Bug Tracker" = "https://github.com/RudraBhaskar9439/Enigma/issues"

[project.scripts]
server = "server.app:main"
inference = "inference:main"

[tool.setuptools]
py-modules = ["app", "inference", "validate"]

[tool.setuptools.packages.find]
include = ["env*", "agents*", "training*", "server*", "swarms*"]
exclude = ["checkpoints*", "logs*", "docs*", "tests*"]
